+++ src/f_finale.c
@@ -753,7 +753,7 @@ static void F_TextWrite(void)
       continue;
     }
               
-    c = M_ToUpper(c) - HU_FONTSTART;
+    c = ST_ToUpper(c) - HU_FONTSTART;
     if (c < 0 || c >= HU_FONTSIZE || hu_font[c] == NULL)
     {
       cx += 4;
@@ -1677,7 +1677,7 @@ static void F_CastPrint(const char* text, int y) // [Nugget] Y parameter
     c = *ch++;
     if (!c)
       break;
-    c = M_ToUpper(c) - HU_FONTSTART;
+    c = ST_ToUpper(c) - HU_FONTSTART;
     if (c < 0 || c >= HU_FONTSIZE || hu_font[c] == NULL)
     {
       width += 4;
@@ -1696,7 +1696,7 @@ static void F_CastPrint(const char* text, int y) // [Nugget] Y parameter
     c = *ch++;
     if (!c)
       break;
-    c = M_ToUpper(c) - HU_FONTSTART;
+    c = ST_ToUpper(c) - HU_FONTSTART;
     if (c < 0 || c >= HU_FONTSIZE || hu_font[c] == NULL)
     {
       cx += 4;