+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## New Features
 
+- **Support for high-resolution sprites between `HI_START`/`HI_END` markers**
 - **DeHackEd thing sprite `Scale` property** support
 - **Weapon Carousel fadeout** setting
 - Toggle to **organize screenshots**
+- **Support for lowercase characters in console font**
 - Toggle to **use palette colors exactly when gamma correction is disabled**
 
 ## Changes